requirements.txt
asgiref==3.5.0
Django==3.2.18
django-cors-headers==3.11.0
django-extensions==3.2.3
django-js-asset==2.1.0
django-mptt==0.14.0
django-mptt-admin==2.5.1
djangorestframework==3.13.1
mssql-django==1.2
Pillow==9.0.1
psycopg2-binary==2.9.6
pyodbc==4.0.39
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
gunicorn==21.2.0